Move verifiers
