+++ components/membership/registration-impl/src/main/kotlin/net/corda/membership/impl/registration/verifiers/P2pEndpointVerifier.kt
@@ -1,4 +1,4 @@
-package net.corda.membership.impl.registration.dynamic.verifiers
+package net.corda.membership.impl.registration.verifiers
 
 import net.corda.membership.lib.MemberInfoExtension
 import java.net.URI
