+++ components/membership/registration-impl/src/main/kotlin/net/corda/membership/impl/registration/verifiers/RegistrationContextCustomFieldsVerifier.kt
@@ -1,4 +1,4 @@
-package net.corda.membership.impl.registration.dynamic.verifiers
+package net.corda.membership.impl.registration.verifiers
 
 import net.corda.membership.lib.MemberInfoExtension.Companion.CUSTOM_KEY_PREFIX
 
@@ -13,8 +13,9 @@ internal class RegistrationContextCustomFieldsVerifier {
     fun verify(context: Map<String, String>): Result {
         val customFields = context.filter { it.key.startsWith(CUSTOM_KEY_PREFIX) }
         if (customFields.size > MAX_CUSTOM_FIELDS ) {
-            return Result.Failure("The number of custom fields (${customFields.size}) in the registration context is larger than " +
-                "the maximum allowed ($MAX_CUSTOM_FIELDS).")
+            return Result.Failure(
+                "The number of custom fields (${customFields.size}) in the registration context is larger than " + "the maximum allowed ($MAX_CUSTOM_FIELDS)."
+            )
         }
 
         var errorMessages = ""
